Исправлена ошибка в irs::modbus_client_t::write_bit, которая приводила к сбросу записанных ранее битов, при вызове функции irs::modbus_client_t::write_bit несколько раз подряд. Добавлены классы irs::const_index_matrix2_t, irs::const_index_matrix_t. В интерфейс irs::mxdata_assembly_t добавлен метод options. Изменен интерфейс irs::mxmultimeter_assembly_t и его реализации. В irs::ini_file_t добавлены перегруженные методы add.
git-svn-id: https://5-10/svn/mxsrclib/trunk@1270 63356120-0428-6b47-bcd6-cd59724df219

+++ arch/cbuilder/tstlan5lib.cpp
@@ -1313,7 +1313,7 @@ void irs::tstlan::view_t::controls_t::bstr_to_var(int a_var_index,
 }
 long double irs::tstlan::view_t::controls_t::var_to_long_double(int a_var_index)
 {
-  double val = 0;
+  long double val = 0;
   netconn_t::item_t item = m_netconn.items[a_var_index];
   switch (item.type) {
     case type_bit: {
